--- src/pages/UserSpeechbubble.jsx
@@ -1,11 +1,0 @@
-const chatData = {
-  id: 1,
-  nickname: "김철수",
-  txt: "ㅋㅋㅋㅋżzzzzzㅋㅋㅋㅋżzzzzzㅋㅋㅋㅋżzzzzzㅋㅋㅋㅋżzzzzzㅋㅋㅋㅋżzzzzzㅋㅋㅋㅋżzzzzz뭐해",
-  profile: "",
-  time: "10시30분",
-};
-const UserSpeechbubble = () => {
-  return <></>;
-};
-export default UserSpeechbubble;
